+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## 2025.11 (June, 6, 2025)
 
+* Fix `pcre_jit` directive context
 * Support lua directives only in Pro version
 
 ## 2025.10 (May, 31, 2025)